# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sistema de Renta de Apartamentos - API",
        "description": "Documentación de la API del sistema de rentas de apartamentos. Permite gestionar usuarios, apartamentos, quejas, contratos y más.",
        "contact": {
            "email": "tucorreo@ejemplo.com"
        },
        "license": {
            "name": "Todos los derechos reservados - Universidad Piloto de Colombia"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/solicitud-alquiler": {
            "get": {
                "tags": [
                    "SolicitudAlquiler"
                ],
                "summary": "Listar todas las solicitudes de alquiler",
                "operationId": "1166f9955daa74cd3ff0266251034002",
                "responses": {
                    "200": {
                        "description": "Lista de solicitudes de alquiler",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/SolicitudAlquiler"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "SolicitudAlquiler"
                ],
                "summary": "Crear una nueva solicitud de alquiler",
                "operationId": "3d4cfe2c470f92bc26c2e7f457719f3b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SolicitudAlquiler"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Solicitud de alquiler creada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SolicitudAlquiler"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
...
        "/api/solicitud-alquiler/{id}": {
            "get": {
                "tags": [
                    "SolicitudAlquiler"
                ],
                "summary": "Mostrar una solicitud de alquiler específica",
                "operationId": "c4925030c065569d76e73fee3043d59c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la solicitud de alquiler",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Solicitud encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SolicitudAlquiler"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Solicitud de alquiler no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "SolicitudAlquiler"
                ],
                "summary": "Actualizar una solicitud de alquiler",
                "operationId": "fef83f23ea98db88a1cc6df1f1904edb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la solicitud",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SolicitudAlquiler"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Solicitud de alquiler actualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SolicitudAlquiler"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Solicitud de alquiler no encontrada"
                    }
                }
            },
            "delete": {
                "tags": [
                    "SolicitudAlquiler"
                ],
                "summary": "Eliminar una solicitud de alquiler",
                "operationId": "c43d2ec0ccd915c293175bc09f543eaa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la solicitud a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Solicitud eliminada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Solicitud de alquiler eliminada correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Solicitud de alquiler no encontrada"
                    }
                }
            }
        },
        "/api/solicitud-alquiler/{id}/edit": {
            "get": {
                "tags": [
                    "SolicitudAlquiler"
                ],
                "summary": "Mostrar formulario para editar una solicitud de alquiler",
                "operationId": "cb181e1a3849e13bc5efa4132e106c03",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la solicitud de alquiler",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Datos de la solicitud para editar",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SolicitudAlquiler"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Solicitud de alquiler no encontrada"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "SolicitudAlquiler": {
                "title": "SolicitudAlquiler",
                "required": [
                    "fecha_solicitud",
                    "usuario_id",
                    "apartamento_id",
                    "estado_solicitud"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true,
                        "example": 1
                    },
                    "fecha_solicitud": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-04-08"
                    },
                    "usuario_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "apartamento_id": {
                        "type": "integer",
                        "example": 3
                    },
                    "estado_solicitud": {
                        "type": "string",
                        "example": "pendiente"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}